fix(dashboard): address policy workspace review feedback
Extract policy table views into a dedicated module, append rule reasons
to all scope summaries, and wire exception sorting through the parent.
