+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/supply-chain-workspace-layout.js
@@ -1,0 +1,4 @@
+const SUPPLY_CHAIN_WORKSPACE_SHELL_CLASS = "min-w-0 max-w-full space-y-6 overflow-x-hidden";
+export {
+  SUPPLY_CHAIN_WORKSPACE_SHELL_CLASS as S
+};
